+++ packages/server/src/error.ts
@@ -1,7 +1,7 @@
 import type { ORPCErrorCode, ORPCErrorOptions } from '@orpc/client'
 import type { ErrorMap, ErrorMapItem, InferSchemaInput } from '@orpc/contract'
 import type { MaybeOptionalOptions } from '@orpc/shared'
-import { fallbackORPCErrorStatus, ORPCError } from '@orpc/client'
+import { ORPCError } from '@orpc/client'
 import { resolveMaybeOptionalOptions } from '@orpc/shared'
 
 export type ORPCErrorConstructorMapItemOptions<TData> = Omit<ORPCErrorOptions<TData>, 'defined' | 'status'>
@@ -45,31 +45,4 @@ export function createORPCErrorConstructorMap<T extends ErrorMap>(errors: T): OR
   })
 
   return proxy as any
-}
-
-export async function validateORPCError(map: ErrorMap, error: ORPCError<any, any>): Promise<ORPCError<string, unknown>> {
-  const { code, status, message, data, cause, defined } = error
-  const config = map?.[error.code]
-
-  if (!config || fallbackORPCErrorStatus(error.code, config.status) !== error.status) {
-    return defined
-      ? new ORPCError(code, { defined: false, status, message, data, cause })
-      : error
-  }
-
-  if (!config.data) {
-    return defined
-      ? error
-      : new ORPCError(code, { defined: true, status, message, data, cause })
-  }
-
-  const validated = await config.data['~standard'].validate(error.data)
-
-  if (validated.issues) {
-    return defined
-      ? new ORPCError(code, { defined: false, status, message, data, cause })
-      : error
-  }
-
-  return new ORPCError(code, { defined: true, status, message, data: validated.value, cause })
 }
