+++ packages/server/src/procedure-client.test.ts
@@ -1,14 +1,19 @@
 import { ORPCError } from '@orpc/client'
+import { validateORPCError } from '@orpc/contract'
 import { HibernationEventIterator } from '@orpc/standard-server'
 import * as z from 'zod'
-import { createORPCErrorConstructorMap, validateORPCError } from './error'
+import { createORPCErrorConstructorMap } from './error'
 import { isLazy, lazy, unlazy } from './lazy'
 import { Procedure } from './procedure'
 import { createProcedureClient } from './procedure-client'
 
-vi.mock('./error', async origin => ({
+vi.mock('@orpc/contract', async origin => ({
   ...await origin(),
   validateORPCError: vi.fn((map, error) => error),
+}))
+
+vi.mock('./error', async origin => ({
+  ...await origin(),
   createORPCErrorConstructorMap: vi.fn(),
 }))
