+++ packages/server/src/procedure-client.ts
@@ -6,11 +6,11 @@ import type { ORPCErrorConstructorMap } from './error'
 import type { Lazyable } from './lazy'
 import type { AnyProcedure, Procedure, ProcedureHandlerOptions } from './procedure'
 import { mapEventIterator, ORPCError } from '@orpc/client'
-import { ValidationError } from '@orpc/contract'
+import { validateORPCError, ValidationError } from '@orpc/contract'
 import { asyncIteratorWithSpan, intercept, isAsyncIteratorObject, resolveMaybeOptionalOptions, runWithSpan, toArray, value } from '@orpc/shared'
 import { HibernationEventIterator } from '@orpc/standard-server'
 import { mergeCurrentContext } from './context'
-import { createORPCErrorConstructorMap, validateORPCError } from './error'
+import { createORPCErrorConstructorMap } from './error'
 import { unlazy } from './lazy'
 import { middlewareOutputFn } from './middleware'
